move route logic to service folder
